pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axprotocol"
version = "2.4.0"
description = "Self-auditing, directive-aware AI reasoning kernel with multi-role agents and TAES evaluation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AxProtocol Team"}
]
keywords = ["ai", "reasoning", "governance", "multi-agent", "taes", "directives"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "PyJWT>=2.8.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pynacl>=1.5.0",  # For Ed25519 signing
    "cryptography>=41.0.0",  # For HMAC fallback and general crypto
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]
ui = [
    "streamlit>=1.28.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
]
vector = [
    "pymilvus>=2.6.0",
    "grpcio>=1.66.0",
    "ujson>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/axprotocol"
Documentation = "https://github.com/yourusername/axprotocol#readme"
Repository = "https://github.com/yourusername/axprotocol"
Issues = "https://github.com/yourusername/axprotocol/issues"

[tool.setuptools]
packages = ["axprotocol", "axp"]
include-package-data = true

[tool.pytest.ini_options]
testpaths = ["verify"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

